Add support to URLRequest for deferring redirects.
I chose to add an out parameter to OnReceivedRedirect because it allows for the
default behavior to remain the same.

I considered adding a ContinueAfterRedirect method that all OnReceivedRedirect
implementations would need to call, but this caused one annoying problem:  In
the case of a ChromePlugin, it is possible for the URLRequest to get deleted
inside the handler for the redirect.  This would make it hard to subsequently
call a method on the URLRequest since I would need to have a way to determine
if the URLRequest had been deleted.

TEST=covered by unit tests
BUG=16413,6442
R=eroman,wtc

Review URL: http://codereview.chromium.org/155897

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21417 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9d73e3f2ca5e1673669a8cd9c6685582d9c0bbfe