Fix NativeViewHostGtk clipping.
Basically, I introduce an additional GtkFixed into the GtkWidget hierarchy. When clipping is required, this is re-created with a X Window and clipping is performed. For some reason, Gtk+Cairo won't clip rendering to the parent GtkWidget if there isn't an associated X Window.

This also fixes the GetPreferredSize implementation for NativeButtonGtk which would return ever increasing values of size each time it was called.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/159153

git-svn-id: http://src.chromium.org/svn/trunk/src@21423 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 160cf61b3259efbc094768213fe3274bae39f1c7