GTK: Don't close info bubbles based on which toplevel is active.
This fixes the behavior for active toplevel windows and focus follows mouse that moving the mouse onto and then off of the bookmark bubble will close it.

This also fixes that we might paint the tab strip active while the system title bar is painted inactive. Unfortunately we can't stop the browser window from painting inactive when using the system frame.

This reverts us to the old bad behavior that you could drag around the browser window and the bookmark bubble will stay in place. It also reverts us to allowing the info bubble to being able to persist even if it doesnt have focus. This doesn't match windows, but I'm not convinced windows is all that great.

BUG=none
TEST=see above

Review URL: http://codereview.chromium.org/159288

git-svn-id: http://src.chromium.org/svn/trunk/src@21576 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 119aad18b2fe4ceadcf76d5bb9db3c669c242db6