Make downloads not prevent tabs from closing.
If a download creates a cross-site transition (for example, if you click a
link in Gmail that results in a download in a new tab), that tab will be stuck
and you can't close it or the browser.

This is the opposite problem with a similar cause as bug 16246. In both cases
we were using some secondary signal to tell us if we're closing for a cross site
transition or closing the tab, and that signal was wrong.

In this case, we were running the onunload handler, but because there was a
pending RenderViewHost, the RenderManager would think that the close was a
cross-site one, and not forward the close message to actually close the tab.

This patch adds a flag to the on unload handlers that indicates whether it's
for a tab closure or a cross-site transition, so we can do the right thing
unambiguously when the message returns. In this case I keep this information in
the RenderView in case we send multiple close requests, we'll close the tab if
any of them were for the entire tab, even if that particular one was dropped
because we don't want to have more than one in flight at once.

BUG=17560
TEST=none.

Review URL: http://codereview.chromium.org/160122

git-svn-id: http://src.chromium.org/svn/trunk/src@21673 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b1b36a0a7c38324f419bad8d957f6060ea208fd5