Quick work around for av_seek_frame() causing video to go completely out of sync.
If we have a freshly opened AVFormatContext, telling it to seek to the beginning (pts=0) apparently causes the video and/or audio streams to get timestampped incorrectly, causing video sync issues.  If the video has been playing a bit, a seek to the beginning causes no issues.  Go figure.

BUG=17832
TEST=ogg videos should remain in sync

Review URL: http://codereview.chromium.org/159438

git-svn-id: http://src.chromium.org/svn/trunk/src@21720 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 29a6beb95f7dc520201ab59552c4cfe6cbc18820