Workaround for GtkFileChooserButton bug: disable mouse wheel scrolling.
Also disable it for the cookie behavior combobox, because it fights with the scrolled window.

The GtkFileChooserButton spews the following 2 errors (and possibly others) when you scroll on it quickly:

Gtk-WARNING **: idle activate multiple times without clearing the folder object first.
Gtk: gtk_tree_model_get_iter: assertion `path != NULL' failed

This causes us to crash in debug, and causes undesirable console output in release. I verified that this is their bug and not ours by reproing in a simple program that just has a window with a GtkFileChooserButton in it.

BUG=17805

Review URL: http://codereview.chromium.org/159459

git-svn-id: http://src.chromium.org/svn/trunk/src@21804 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5fa3612c92f7abc72323535103e68a40147e754f