Fixes an audio playback regression caused by seeking in audio/video media.
This is a temporary fix to a much larger problem of dealing with Chrome's audio IPC layer.  Since a seek is a small amount of time, we keep the audio IPC conversation going by writing 8k of zeros.  Before we were telling the audio IPC layer that we were out of data, which technically was a lie (we were just seeking), and as a result the conversation died at that point.

TEST=seek around in a video, audio should keep playing and remain in sync
BUG=17917

Review URL: http://codereview.chromium.org/160283

git-svn-id: http://src.chromium.org/svn/trunk/src@21935 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 728e509783f8726e84902e72122c278cc77836a5