Retry requests that has timed out in BufferedDataSource
Retries request if the read request is not fulfilled in 5 seconds and at most 3 times. This will handle the case when
a user pause a movie and starts it in a later time when the
connection to the server has already timed out.

TEST=BufferedDataSourceTest.*
BUG=17479

done

Review URL: http://codereview.chromium.org/159382

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21982 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 09fff1f3588f7345299c7851ada580caf50ba46b