Removing crash 17710 as the underlying issue has been fixed in upstream webkit.
Adding two flash crashes to the known crashes list.

These crashes seem to be relatively infrequent so it should be safe to add these to the known crashes list for now, while we investigate the underlying cause.

The associated bugs are http://code.google.com/p/chromium/issues/detail?id=18058
and http://code.google.com/p/chromium/issues/detail?id=18059

Both crashes occur in the context of an incoming NPP_Write call to the plugin while it is an outgoing NPN_Invoke call.

I will debug these issues in parallel as these started occuring as a result of r22041 which was to make the plugin stream IPCs async which in my opinion is the right fix.

I will revert 22041 if similar crashes prop in the next day or two.

TBR=huanr


Review URL: http://codereview.chromium.org/159630

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@22046 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 21e56c62acc10dc9f8f4cec13993c1e54bc92239