Missing buffered attribute for <video>/<audio>
BUG=16056
TEST=LayoutTests/media/video-buffered.html

The current implementation of the buffered attribute for <video> and <audio>
is broken. There are several problems around this attribute:

1. We don't have any caching on disk, so we only keep a recent range of bytes
   close to current playback position.
2. WebKit reports buffered as one range (0, max_time_buffered). But we only
   cache a short partial range which doesn't start with 0. The correct
   implementation is a list of ranges buffered. But this has to go into WebKit
   first.
3. We don't have an accurate mapping between byte offset < - > timestamp.

So the current implementation is to lie about what we have buffered. We always
say we have buffered everything before the current download position. And we
only report one range. The calculation of time is also based on scaling the
duration with current buffered bytes.

Review URL: http://codereview.chromium.org/160300

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@22087 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a000230a08bff6fa9c472f03961ccb536f0239aa