Linux: Fix Valgrind error in VisitedLinkEventsTest.Coalescense.
I introduced the error in r22193; it shows up when there's no X
server to connect to.  I get a "GLib-GObject-CRITICAL **:
g_object_get: assertion `G_IS_OBJECT (object)' failed" message with
this change, but I'm not sure if it's from my code or if it was
already there, and the tests still pass.

TEST=ran VisitedLinkEventsTest.Coalescense under Valgrind with DISPLAY=:6

Review URL: http://codereview.chromium.org/159753

git-svn-id: http://src.chromium.org/svn/trunk/src@22240 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 541e4f591d92d378224b685138a1fa97655e80f0