GTK Themes: Theme labels in LocationBarViewGtk.
Labels in LocationBarViewGtk should be black even when a dark GTK theme is used
when we're not in GTK theme mode. Likewise, when we're using a GTK theme, we
should have the background and border color of the "Search X for:" box use
colors from the theme instead of having hard coded ones.

Review URL: http://codereview.chromium.org/159798

git-svn-id: http://src.chromium.org/svn/trunk/src@22293 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b7c32cf2f7b62d936c9e678eb1330cc47308e27a