Implement a couple layoutTestController methods and rebaseline http/tests/misc/favicon-loads-with-images-disabled.html.
Implement layoutTestController.disableImageLoading() and put
in a placeholder for layoutTestController.setIconDatabaseEnabled()
(because we don't use the icon database).

These methods are used by
http/tests/misc/favicon-loads-with-images-disabled.html, but the
test still fails because our resource loader output is different.
Go ahead and rebaseline for the difference.

BUG=18355

Review URL: http://codereview.chromium.org/159820


git-svn-id: http://src.chromium.org/svn/trunk/src@22345 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5668461d1c2a58fea72eaff0539f7b365e17ac42