Split out HttpUtil::SpecForRequest() into a more generic function of net_util.h.
This was a TODO, since that function is useful outside of HTTP.

In the process, I uncovered some test cases in proxy_service that are passing in invalid URLs (by virtue of the extra DCHECK). This doesn't make much sense to me to support that, so I have changed them.

Review URL: http://codereview.chromium.org/160558

git-svn-id: http://src.chromium.org/svn/trunk/src@22359 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c01b3188fff313cbca3953b0610cca5d9aba9bcb