Don't call NPP_SetWindow during the painting of windowless plugins.
On Windows, Flash seems to only start executing script actions after it received an NPP_SetWindow with a 
non-NULL NPWindow.window (HDC). It is possible that Flash then invokes JS to modify DOM of the page.

If Flash movie's widget is on-screen at page load, this call is made during layout and before even the NPP_Write is called,
which is the desired sequence of events.

However, if it is off-screen, this call occurs during painting, which leads to re-entrancy issues (layout while painting)
and bizarre crashes.

As a solution, we remove calls to NPP_SetWindow during painting and instead opt to never provide a null HDC to the plugin.
If no valid HDC is available, we feed it a disposable monochrome 1x1 context to have at least something to draw on.

R=ananta,darin,jam
BUG=16114
TEST=LayoutTests/plugins/flash-setwindow-paint-crash.html (bug reduction).

Review URL: http://codereview.chromium.org/159717

git-svn-id: http://src.chromium.org/svn/trunk/src@22383 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f62afe124054c4f7f4d8ac4dd308c2455d8b0473