This CL is doing a bunch of Misc work to make LocalStorage data persist.
First of all, this allows WebKit clients to specify whether or not the VFS should be used.  In the browser process, we never want it to be.

Next, this allows WebKit clients to specify the behavior of WebKit's FileSystem code.  By default, they should all be NOT_REACHED().  The browser process implements the two of these I need for LocalStorage, but it'll be really easy for the rest to be implemented as needed.

Next, this adds a function that storage routines can call to ensure that lazily initialized stuff that must be initialized on the main WebKit thread is initialized.  Right now, this is just used to initialize the UTF8 conversion tables, but I'm sure there'll be more overt time.

Lastly, this uses the profile directory stored by webkit_context_ to derive the path LocalStorage should use.

This CL also cleans up a few minor style issues as it goes.

TEST=none
BUG=4360

Review URL: http://codereview.chromium.org/159778

git-svn-id: http://src.chromium.org/svn/trunk/src@22452 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2f44f380dca0e97c3a8fc56454694ecc8ba0ac10