Run the DOM Storage layout tests in the UI test framework since the test shell doesn't hit very much of the new code.  I still plan to write unit tests and we're still goign to enable the layout tests in test shell, but this is still a good way for us to test the full stack ASAP.
Workers has used this technique for a while now.  This CL factors out the layout test running code from the workers ui test and moves arounds directories a bit so not everything is in the "workers" directory.

Doing this via a v8 extension is the right way to go long term, but I ran into a couple snags, so I think this is a good first step.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/159720

git-svn-id: http://src.chromium.org/svn/trunk/src@22466 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 817b5e357b8398a4e067793281720fc5e086de21