Ensure that we never call into WebCore::Page static methods when Page constructor hasn't been called.
This is a top crash for Mac right now. The crash occurs because the static table of all pages in the renderer is initialized 
in WebCore:: Page constructor, and if we attempt to access this table before the Page was created (i.e. before WebView::Create is called),
we get into a crashy situation. I am really not sure how to reproduce this, but I can defend against the situation.

BUG=17555
R=darin
TEST=unable to repro, speculative fix.

Review URL: http://codereview.chromium.org/159887

git-svn-id: http://src.chromium.org/svn/trunk/src@22540 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8654dcd535ff7f6ad0cb5ea7b550a5b3cf4238ae