A step in exposing Bitmap to JavaScript.
The plan is to make pack.createBitmapFromRawData
return an array of bitmaps.

1 bitmap if it's a standard 2d image
6 bitmaps if it's a cubemap
N bitmaps if it's a volume map.

The bitmaps will have a semantic so you can look at them and tell
they were from a volumemap, a cubemap or a 2d image.

On the way I'm attempting to clean up the code. Moving
stuff out of Bitmap the did not belong there and 
Refactoring Bitmap so there are less If Format =
stuff.

Review URL: http://codereview.chromium.org/164034

git-svn-id: http://src.chromium.org/svn/trunk/src@22583 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 02072e48e15e7771c8e230195caa0e435d866c9d