Fix some issues with bookmark bar folder menu.
Without this patch, you have to click repeatedly on a folder to get the context menu to pop up, because the first few clicks dismiss whatever submenus you may have open. This patch fixes that.

There is another problem present with and without this patch where a bookmark bar folder menu that has shown a context menu will not regain the application/X grabs. This makes it possible to use the rest of the browser while the menu is showing, and clicking away from the menu won't dismiss the menu. In the interests of keeping this patch easy to review I will look into that issue separately.

BUG=18275

Review URL: http://codereview.chromium.org/165138

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@22743 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f173d925c209a5c8b4fa95d460c7f0197e9a1e06