Fix for an "update in progress" DCHECK.
A DCHECK in SafeBrowsingService::UpdateStarted can potentially be
triggered if there is an error opening the database for reading,
or if there is no response from the server for a update request.
Both of these cases can cause our 'update_in_progress_' flag to
not be reset properly before the next update.

This CL adds:
- a timeout for update responses
- better handling for database errors
that properly resets the update state.

BUG=12835 (http://crbug.com/12835)
TEST=None.

Review URL: http://codereview.chromium.org/165008

git-svn-id: http://src.chromium.org/svn/trunk/src@22814 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1be808a25743ba59436b21b74e1782b910de5417