This CL adds client.toDataURL which gets the contents of the client area as a data url (which is a base64 encoded string of a png file)
data urls are part of the HTML5 standard and supported
by firefox, safari and chrome.
http://www.whatwg.org/specs/web-apps/current-work/multipage/the-canvas-element.html#dom-canvas-todataurl

That means you can now do this

var dataURL = client.toDataURL();
// make an IMG tag display the screenshot
someImgTag.src = dataURL;  
// use the IMG tag to draw into a canvas
someCanvasContext.drawImage(someImageTag, ...);

It also means there is no need for the test builds
anymore "test-dbg-d3d", "test-opt-d3d" etc as
toDataURL is part of the public API and can
therefore always be used to take screenshots
in any build.

I updated the selenium code to use it.

There are a few issues:

1) The GL version has the same limitations as taking
a screenshot before. Namely that the client area
must be on screen. We need to fix this.

2) We need to add support for origin-clean. See
https://tracker.corp.google.com/story/show/180334

Review URL: http://codereview.chromium.org/164130

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@22869 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c20890387ffb8b22ccc12348fcd0ad525dbe5343