Rework visited link updating mechanism to be more robust.
This is a follow-up to http://src.chromium.org/viewvc/chrome?view=rev&revision=22540,
which eliminated one crash, but still left behind another, this time due to
racing between history backend and WebView creation.

The solution is to make sure no visited-link-related messages are sent to the
rendering process until we know for sure it is created.

BUG=17555
TEST=VisitedLinkeRelayTest.WebViewReadiness
R=darin


Review URL: http://codereview.chromium.org/165210

git-svn-id: http://src.chromium.org/svn/trunk/src@22917 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b04b581a4d915d80f29df7585cbc802ccc6b4b3e