Rework visited link updating mechanism to be more robust.
This is a follow-up to http://src.chromium.org/viewvc/chrome?view=rev&revision=22540,
which eliminated one crash, but still left behind another, this time due to
racing between history backend and WebView creation.

The solution is to make sure no visited-link-related messages are sent to the
rendering process until we know for sure it is created.

BUG=17555
TEST=VisitedLinkeRelayTest.WebViewReadiness
R=darin


Review URL: http://codereview.chromium.org/165210

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@22917 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7b7f68dd13fb0b73276c20b54e5727bc79141196