linux: fix race in TransportDIB passing
There was a race condition when the renderer passes a TransportDIB to the plugin
process. It is possible that the renderer destroys the TransportDIB before the
plugin handles the message, in which case the plugin will fail to map the shm
handle.

This also cleans up a couple of things.

BUG=18320

Review URL: http://codereview.chromium.org/164226


git-svn-id: http://src.chromium.org/svn/trunk/src@22949 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4da103873abe6e9827517e3a3f4f533c2e6f6bb8