Commit patch for Pierre-Antoine LaFayette, original CL: http://codereview.chromium.org/164119
Items in the Downloads page should be loaded like a normal OS file when dragged
to a Tab content area. We need the URL to be set to the file path in the
IDataObject so that WebDropData can properly handle the file.

BUG=9266
TEST=Drag an item with a supported MIME type from the Downloads page to a Tab
content area. It should load within the browser. Drag an item with an
unsupported MIME type and it should not open it.

Review URL: http://codereview.chromium.org/165260

git-svn-id: http://src.chromium.org/svn/trunk/src@22983 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: be62dee4477f3c40f37102368f66c4864e7361b7