I have taken over this patch, original authored by Adam Hunter (adamhunter).
The original issue is: http://codereview.chromium.org/144019

--------------

Functionality has been requested in the Extension API for Javascript to
take screenshots of the currently visible tab.  This changelist builds this
function, chrome.tabs.getVisibleScreenCapture.  This function takes a
single callback function and returns to that function a data URL of a JPEG
image of the current screen.  A simple sample extension is provided as a
use case.

BUG=14760
TEST=There is an extension in chrome\test\data\extensions\samples\screenshot,
load this extension.  It creates a toolstrip button.  Click this button, you
should get a page with a screenshot of the active tab.  The API function
is found at chrome.tabs.getVisibleScreenCapture.

Review URL: http://codereview.chromium.org/160228

git-svn-id: http://src.chromium.org/svn/trunk/src@23259 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 68898587e0b9cc2e308b620d1942f8bcc54018d4