Add an UpdateContext call to WebPluginDelegateImpl on the Mac.
On the Mac, Flash appears to cache the CGContextRef provided in NPP_SetWindow until the next NPP_SetWindow call, so we need to call it sometime before the next plugin paint. This allows us to call NPP_SetWindow before telling the plugin to paint, but not from the Paint function itself (where it could have bad side-effects).

BUG=18894,18980
TEST=Switch to HQ on a YouTube video; Flash should not crash.

Review URL: http://codereview.chromium.org/165344

git-svn-id: http://src.chromium.org/svn/trunk/src@23322 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e303d8675e7632a6b6bf2e5742349842cfc3813a