Restore the WebFrame::getPrintPageShrink() method originally added in rev 14639 and later removed in rev 19669. From the original commit message:
Rev 12100 changed the way that we spool and render printed output, most notably
introducing use of the PrintContext class.  The existing
PrintContext::spoolPage() method applies a webkit scaling factor before
rendering output to the graphics context.  ChromePrintContext::spoolPage() (in
webframe_impl.cc), which is used by chromium instead of
PrintContext::spoolPage(), does not apply this scaling factor, but instead
eventually returns the scaling factor via WebFrame::PrintPage().  This is a
problem for the Chromium Embedded Framework (CEF) because, unlike chromium, the
CEF renders directly to the printer device context.  It is therefore important
for CEF that we retrieve and apply the webkit scaling factor before calling
PrintPage().  In order to support this capability the following adds a
WebFrame::GetPrintPageShrink() method.

Patch by Marshall Greenblatt  
R=darin
BUG=none
TEST=none



git-svn-id: http://src.chromium.org/svn/trunk/src@23338 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 707d67cb8d409464d1aea24151a8e89d1d40d92b