When you clicked Get Themes in Mac prefs, the browser window containing the themes gallery wasn't coming to the front.
There was an implicit assumption on Windows that TabContentsViewWin::Focus caused the window containing the TabContents to be foregrounded. This is because on Windows a HWND is focused with a call to SetFocus, which activates the containing top level window. On Mac, TabContentsViewMac::Focus needs to explicitly activate the containing window.

Also, changes the Mac preferences code to call the same function as on Windows, for simplicity.

Note: the "foreground" parameter to AddTabWithURL means the tab is opened in the foreground, not that the window is activated.

http://crbug.com/18157
TEST=none

Review URL: http://codereview.chromium.org/165492

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23378 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 720b94e1d6213f850fdf200a848d925db2ce5549