Change to O3D IMC code so UPDATE_TEXTURE2D_RECT has a pitch which allows you to send a sub rect of a source rectangle.
Also fixed a bug in UPDATE_TEXTRE2D so it works
as it used to. They used to send 1280x199+200 bytes
to update the first 320x200 area of a 1280x720 texture.
I broke that in a previous build.

Finally I changed the declarations in message_command.h
to be C++ compatible.  The offsetof macro does not work if
a class has a constructor.

Review URL: http://codereview.chromium.org/165503

git-svn-id: http://src.chromium.org/svn/trunk/src@23402 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ad551bc72973238660b336150df0b01635805790