Fix typo from r23272.
Also change NavigateToURLLogResult to take a std::string for its URL argument.

Also prepend g_ on all the file-global static variables. This should help prevent future mistakes of a similar nature. This is not really within style but sort of matches the g_ for global variables. Anyway this file is pretty far out of style to begin with.

Review URL: http://codereview.chromium.org/169009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23484 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 88dd58ba7d8dc118df0d56fe816bf1a0e4b848a1