Change selenium to run the samples from the artifacts directory.
Note: There is the issue that the scons build does not
have the correct dependencies to make sure the samples
get copied to the artifacts before selenium runs.

If that's a 1 line fix then let's add it. Otherwise
we can fix that in the gyp build?

Review URL: http://codereview.chromium.org/172008

git-svn-id: http://src.chromium.org/svn/trunk/src@23565 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f5b08ed12b94ea4efc75de36301a4810ee3ae264