A WebApplicationCacheHost api change allow 'foreign' entries to restart the current frame navigation.
If a document has a <html manifest='foo'> attribute that doesn't correspond with the manifest url of the cache
from which the doc was loaded, thats a foreign entry. The entry should be marked as such, and the frame
navigation restarted. The next time around, the foreign entry won't be picked up (because its foreign).

TEST=none
BUG=none


Review URL: http://codereview.chromium.org/164570

git-svn-id: http://src.chromium.org/svn/trunk/src@23655 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f66bcf09394edd26538ba6d5f30ef9bee348d40c