Linux: Revert updating PRIMARY selection on Ctrl-C in omnibox.
This backs out r23596.  I'm pretty sure that it worked when I tested it
right before submitting, but it's totally breaking stuff now -- it looks
like when I update the PRIMARY selection, GTK notices that it's lost the
selection and automatically unhighlights the text in the omnibox, making
it impossible to use Ctrl-C to copy omnibox text to the clipboard. :-(

BUG=19648
TESTED=Hit Ctrl-L, Ctrl-C and confirmed that text remains highlighted and becomes CLIPBOARD selection.

Review URL: http://codereview.chromium.org/174058

git-svn-id: http://src.chromium.org/svn/trunk/src@23714 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8f98b04ef75a0795adf2846d0388777b28e4e514