[Chrome OS] Adds support for injecting Corp cookies at startup
To support single-sign-on for Chrome OS, we need a way to inject cookies into Chrome.  In the case of session cookies, putting them into Chrome's cookie jar DB doesn't work.  This CL adds a command line flag that tells chrome the name of a Unix pipe to open, from which it can read said cookies.

Eventually, I want to replace this pipe-reading with an appropriate usage of DBus, but Chrome OS isn't there yet.  This CL adds the appropriate infrastructure, though, and the PipeReader class can later be replaced with something that pulls the cookies off DBus instead.
Review URL: http://codereview.chromium.org/174062

git-svn-id: http://src.chromium.org/svn/trunk/src@23936 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4007f96bdf55c245d5c1397728cde3df436be1cf