Threadsafe access to AudioSource callback. Release stream on close. Handle error on open better.
This should fix the memory leak on stream close by synchronizing the removal
of the audio source callback with the signaling of a close event to both the
audio source, and to the audio manager.

This change also makes it possible to go from kPlaying -> kPlaying because
the audio source callback can be swapped correctly.

Lastly, Open() is made more robust by failing early on an error state.

BUG=19860,18217
TEST=none

Review URL: http://codereview.chromium.org/173168

git-svn-id: http://src.chromium.org/svn/trunk/src@23988 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c05586031409acc11435b1d866916f3a92f19771