Load extensions that have their manifest in the Prefs on the UI thread.
There is some migration code for loading old extensions which we do on the UI
thread as well, for simplicity. This will only happen the first time the user
starts Chrome, so I think it's OK.

Also cleaned up the EXTENSIONS_LOADED notification to send out a single
extension, instead of a 1-element list of extensions.

BUG=18293
TEST=loading extensions should still work

Review URL: http://codereview.chromium.org/174208

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23997 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4569f4458759330c17c24b703022a12693174c98