linux: call g_thread_init() at relevant startup points
According to glib docs, we need to do this if it's at all possible
for us to hit glib on multiple threads.  This may be happening
when we grab plugin metadata from the file thread.

(Note that this is *not* initializing the GDK locking system.)

BUG=18957

Review URL: http://codereview.chromium.org/174264

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@24043 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a509ba480c80942aff2e0de61190b7e3e5c1c022