Providing unplayed bytes in audio hardware buffer
BUG=20007
TEST=WinAudioTest.PCMWaveStreamPendingBytes

We used to guess what is in the hardware audio buffer to
perform audio sync. But due the triple buffering and
other platforms has other buffering scheme we need to be
able to know how many bytes are left in the hardware buffer.

We provide this data via AudioCallback::OnMoreData(), so we
can predict when the requested buffer will be played.

Review URL: http://codereview.chromium.org/174243

git-svn-id: http://src.chromium.org/svn/trunk/src@24114 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3d1cfdc7468964942ae245d8e6ebdf3e2657a0ad