Linux: exit render processes with the correct code.
When using a zygote, we don't set the |rv| variable in ChromeMain.
Thus, renderers exit with a code of 255. There appears to be a
harmless race where, when closing a render process, sometimes we'll
see the close event from the IPC channel before we believe that the
process has exited. In that case, the 255 exit code makes us believe
that the renderer crashed.

This is probably the cause of the huge renderer crash numbers in UMA.

BUG=20149
http://codereview.chromium.org/173302


git-svn-id: http://src.chromium.org/svn/trunk/src@24157 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7136cd6de7f4d34cfbf530bd868d5dfeba920f6b