Fix "crashed extension" infobar browser crashes.
This is a general rework of how "crashed extension" infobar works
and how the extension is actually recovered after the crash.

This is a subset of original http://codereview.chromium.org/164151/ .
I had to remove the part which unloads the entire extension on crash,
because it interacts badly with other parts of the browser. I'm fixing that.

TEST=See bug.
http://crbug.com/15888

Review URL: http://codereview.chromium.org/173314

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@24231 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a21f13f2dc8aec02f0486aa0df4c9c666e9133e2