Relanding with v8 fix.
Modify the RenderThread to track the number of widgets
and "hidden widgets" which are running through that thread.

By knowing the if the widgets are all hidden, the thread
can accurately inform V8 when it is idle so that V8 can
better cleanup unused memory when idle.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/173466

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@24448 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9a917515a2176169e2c5715ff8e5a7919f57e773