Relanding with v8 fix.
Modify the RenderThread to track the number of widgets
and "hidden widgets" which are running through that thread.

By knowing the if the widgets are all hidden, the thread
can accurately inform V8 when it is idle so that V8 can
better cleanup unused memory when idle.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/173466

git-svn-id: http://src.chromium.org/svn/trunk/src@24448 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: aa25462e599510ecec3d3ec099c7d6be1cd2e206