Don't use URL-encoded strings when user's input is in progress or the text isn't HTTP/HTTPs.
It seems that Firefox is doing the same thing.

BUG=16534 (http://crbug.com/16534)

TEST=Visit any website and append double-quote (") for the URL and invoke
"Select All" by Control-A, then copy the URL by Control-C. Paste the result into
notepad and check if the double-quote is NOT encoded.
TEST=Access file:///" and select whole Omnibox and copy the URL. Paste the text
into notepad and check if the double-quote is NOT URL encoded.
TEST=Put " in Omnibox and type enter key so that you will be navigated to
http://www.google.co.jp/search?sourceid=chrome&ie=UTF-8&q=%22. Copy whole
content in Omnibox and paste it into notepad. Then, check if the double-quote is
URL-encoded (%22).

Original Review: http://codereview.chromium.org/155495/show
Patch by hamaji@chromium.org (he's on vacation and asked me to land) 



git-svn-id: svn://svn.chromium.org/chrome/trunk/src@24472 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d1beac1457a060290b7bd66625c9d8e4d0ccebb3