Don't use URL-encoded strings when user's input is in progress or the text isn't HTTP/HTTPs.
It seems that Firefox is doing the same thing.

BUG=16534 (http://crbug.com/16534)

TEST=Visit any website and append double-quote (") for the URL and invoke
"Select All" by Control-A, then copy the URL by Control-C. Paste the result into
notepad and check if the double-quote is NOT encoded.
TEST=Access file:///" and select whole Omnibox and copy the URL. Paste the text
into notepad and check if the double-quote is NOT URL encoded.
TEST=Put " in Omnibox and type enter key so that you will be navigated to
http://www.google.co.jp/search?sourceid=chrome&ie=UTF-8&q=%22. Copy whole
content in Omnibox and paste it into notepad. Then, check if the double-quote is
URL-encoded (%22).

Original Review: http://codereview.chromium.org/155495/show
Patch by hamaji@chromium.org (he's on vacation and asked me to land) 



git-svn-id: http://src.chromium.org/svn/trunk/src@24472 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 67246d22d3eb43d4ba72e418889feb52c221234b