Adds a RENDER message to the O3D IMC.
Normally O3D renders as fast as it can but you can 
put O3D in a render-on-demand mode
in which case it will only render when it has to,
like when part of the window is uncovered, or when
you explicit tell it.

Asking it to render does not render immediately, it
just marks the plugin in a state of "need to render".
The plugin will render at the next available chance.


Review URL: http://codereview.chromium.org/174465

git-svn-id: http://src.chromium.org/svn/trunk/src@24531 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: de9c2869fe99e8e0494226743d303797303b6e1a