Delay calling NotifyEnded() until the audio hardware has played the last of the audio.
Previously we were calling NotifyEnded() when we had sent the last of the audio to the hardware.  This is a small change to call NotifyEnded() when the audio hardware replies asking for additional data.  Not perfect, but better than what we were doing before.

TEST=play a short seamless audio clip on loop, it should be closer to seamless looping
BUG=20349

Review URL: http://codereview.chromium.org/173520


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@24637 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1ae329ac3954a05bafec43b07a486f4221c7ff96