Squash the bookmark bar into the toolbar by 4 pixels.
On windows they do this by having the two views overlap. We can't so instead move the bottom 4 pixels of padding from the toolbar widget into the bookmark bar widget. When the bookmark bar is closed/hidden, it still shows those 4 pixels. When it opens, it is able to make use of those 4 pixels. Since we get 4 extra pixels, we can make the overall height of the bookmark bar 4 pixels less.

BUG=18570

Review URL: http://codereview.chromium.org/178029

git-svn-id: http://src.chromium.org/svn/trunk/src@25075 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6d6ffbce2cf4d41778ceba5ed047d234b993f1ff