Workaround Skia and Pango measuring font heights differently by asking Pango for the height that we'll report in the gfx::Font. This seems wrong, but at least things will look reasonable until we find a better solution.
This changes the existing height computation code to use ceiling. Even though
this is no longer used, it is good to have the correct code for the height
computation here so that if we find a better solution, it will be run properly.

This also gives more room for the clock in the status area view because the
height is now more and we want it to be aligned properly.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/183044

git-svn-id: http://src.chromium.org/svn/trunk/src@25131 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e3834c4b9689c57f7103c0feb4ea27aa7ceaf313