Improve New Tab Page load performance.
This change nets another 65ms (20%) improvement in the duration between the
renderer requesting the NTP and it having all of the resources for the NTP
available and being completely done with layout on my Mac laptop in release
mode.

Move the handling of important resource requests needed for NTP layout from the
UI thread onto the IO thread.  This is handled by caching the necessary data on
the UI thread when possible, so that no profile or theme access is needed to
service the request.  There is often a lot of contention for the UI thread when
a new tab is created, and allowing these requests to be handled more directly
by the IO thread allows the renderer to lay the page out while the browser is
still busy doing UI work.

http://groups.google.com/group/chromium-dev/browse_thread/thread/7148074f807dc5f7

TEST=Do the new tab page and new incognito tab page still work?
     Are they faster?
BUG=13337
Review URL: http://codereview.chromium.org/179069

git-svn-id: http://src.chromium.org/svn/trunk/src@25232 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5afa7fa3dcc5a50f3335abc5a93243dd54eb6d7d