This changes the selenium driver script to not be so scons-specific, and makes it so that it will self-locate some of the things it needs (for instance, it no longer needs to have PYTHONPATH env var set correctly -- it'll find the third party dir relative to itself and add gflags to the path).
As part of this, it now copies the selenium tests into the artifacts directory
so it can run them out of that directory instead of from the source dir.  This
is so that we can use the product dir as the root of the http server, since
in GYP all three platforms have different locations for their output, and
different working directories during the build.

Review URL: http://codereview.chromium.org/180074

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@25288 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ac82a8191ab51d814d86d35685801bb9c9ba46f9