Check the right Info.plist to determine whether the app is the browser or a background helper.
This fixes the Breakpad about:crash problem, but not directly.  I'm not
exactly sure of the failure mechanism here.  It's likely that doing this call
properly here caches the Info.plist dictionary, and that something in Breakpad
crashes when that's not done.  Without this fix, we're seeing a double fault,
with the second exception occurring in the renderer while it is processing
the crash.  Likely because of the use of fork && execv instead of posix_spawn
to launch the renderer, the second exception is delivered to the browser
process.  The browser's Breakpad instance recognizes that the exception is not
its own, but attempts to act on it in part anyway.  There may be a Breakpad
bug here, I'll investigate more tomorrow.

BUG=20371
TEST=In a Breakpad-enabled build, about:crash should bring up a sad tab and
     not crash the browser process
Review URL: http://codereview.chromium.org/194007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@25311 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: da37eb7d9d06c199560521b91caf3beb6b75c65d