(Mac) Make the Page/Wrench buttons place their menu in the right place.
Also makes the buttons drag-able -- you can drag or click and wait to
open their menu. Lucky for me, I wrote reusable code.

To do (not now): The "off-the-end" button on the bookmark bar needs the
same fix.

Bug (not to be fixed now): The menu doesn't do the right thing when the
window is moved off the left edge to the screen -- it doesn't open to
the right in the correct fashion. Nor does it do the proper thing when
you move the window off the bottom edge of the screen -- it doesn't open
upwards entirely correctly. <shrug>

BUG=18572
TEST=Click on the Page and Wrench menu buttons and observe. Also test \
     dragging to open.

Review URL: http://codereview.chromium.org/174556

git-svn-id: http://src.chromium.org/svn/trunk/src@25362 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c2e0dbc2a45612c489e9de6dcaa8aeedea8b4fe4