linux: improve support for cross-compiling
This CL adds support for a 'sysroot' GYP define, that should point to the target root filesystem for cross-compilation.
It passes that argument to the compiler and linker which uses it to prefix its hard-coded path (e.g. /usr/include)
It also points pkg-config to look for package configs there, and rewrite the paths to be prefixed by 'sysroot' (since pkg-config doesn't do it itself)

Review URL: http://codereview.chromium.org/199016


git-svn-id: http://src.chromium.org/svn/trunk/src@25418 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 38a607f8b99f3d724b7e7feb2e4f671f81677097