Make O3D always clear the backbuffer one way or another.
The issue is that given that we are using DISCARD
on D3D and on GL we are only drawing to the visible
portion of the O3D area, the state of the backbuffer
is undefined.  This change should make it consistent
across platforms.

Review URL: http://codereview.chromium.org/201013

git-svn-id: http://src.chromium.org/svn/trunk/src@25522 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ca433c318a0fbe8417eaf120d0a75810b9167eec