Re-enabled URLRequestTestHTTP.BasicAuth.
The problem was the test was expecting |response_time| of cached responses to be the timestamp of the original request, rather than the timestamp of the last cache validation.

Also updated the header to clarify what |request_time| and |response_time| correspond to in the cached cases.

BUG=http://crbug.com/20594

Review URL: http://codereview.chromium.org/197030

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@25540 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b49b7a8c1599a8f1bc0d3f030c9f26b640dc6cb6