Allow the default browser check to return "unknown" and reflect that in the UI.
On Linux this can happen for unsupported desktop environments. On Windows and Mac OS X this probably should not occur and likely indicates some sort of serious configuration error. This change avoids repeatedly bothering the user to set the default browser in cases where we're likely to fail at that, without incorrectly displaying that we are the default browser.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/200025

git-svn-id: http://src.chromium.org/svn/trunk/src@25553 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e783e936bed849f2df83840ce67634eb03df3eae