POSIX: Have pref service save paths as UTF8.
We were passing the raw path string to StringValue, but
StringValue expects strings to be UTF8.  There's only
a DCHECK to enforce this, so I imagine it was working
silently before.

This will break for users that don't currently have UTF8
file names, but it will now work going forward.

Review URL: http://codereview.chromium.org/201027


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@25555 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e5f853ea3d7b6b1b6960e1a54bdf3cb822c9c9f8