POSIX: Have pref service save paths as UTF8.
We were passing the raw path string to StringValue, but
StringValue expects strings to be UTF8.  There's only
a DCHECK to enforce this, so I imagine it was working
silently before.

This will break for users that don't currently have UTF8
file names, but it will now work going forward.

Review URL: http://codereview.chromium.org/201027


git-svn-id: http://src.chromium.org/svn/trunk/src@25555 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3cb6b1ad61f37f0bb3d4484322c0deb064a1a4e0