ForceTLS: hash hostnames, handle subdomains, canonicalise.
It turns out that JSON[Reader|Writer] cannot handle periods in key
names(!). Because of this, an also to avoid leaking a sort of ForceTLS
browser history in the state file, we hash the domain names.

Also, this patch tries to implement the RFCs with respect to
canonicalising the names. Since IDN processing has already occured by
the time the name reaches us, there's only so much that we can do
however.

http://codereview.chromium.org/201033


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@25696 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 805320eb78a30693eb0399957b6bf54ad3b9f388