Add a new wrapper for sqlite. This is mostly a large cleanup of the existing one, combined with the statement cache in a nice way. It is designed to entirely wrap sqlite so that we can catch corrupt errors in the future and "do something" when we get them without having to change all the calling code.
There is also a new meta_table file which is almost exactly like the old one
but which uses the new sql interface.

This patch changes Chrome's history TextDatabase to use this new wrapper as a
proof of concept, because this usage is relatively well-confined.
Review URL: http://codereview.chromium.org/199047

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26022 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d9626ac1e6dba53d20bf9c0c43ee236b140dffa7