Potential fix for http://crbug.com/13111
The browser_window_controller is destroyed before the window is destroyed.
In its dealloc, it sends the download shelf an |exiting| message that deletes
all item controllers, but the items's views might still be visible and hold
weak references to their controllers. When they are clicked, they would se
nd a message to free'd memory.

This is all speculation.

BUG=13111
TEST=Release to dev channel, monitor crash reports. Should go down.

Review URL: http://codereview.chromium.org/193109

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26198 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 09d013544db1895aca29fdb80cb42821a0b6400c