ExtensionShelf now uses the BookmarkExtensionBackground, just like the BookmarkBarView.
Changed the WebKit API to add an optional |id| parameter to the insertStyleText, which is needed to be able to replace style sheets that have been previously added. 
Added an interface that both BookmarkBarView and ExtensionShelf implement. 
This new interface tells us whether we are located at the top or at the bottom and whether we are detached from the frame or not. 
Factored out some of the duplicate painting-related code to a namespace of its own. Not happy with the name (welcome suggestions). 
Moved the check for whether extensions are on top to new class and now cache the value for the lifetime of the process. 
Toolstrip text color values are no longer hard-coded but use the color specified in the theme. 
Decreased the timeouts for showing and hiding the toolstrip handle. 
Replaced the pressed background image and the hover background image for the toolstrip to match what the bookmark bar uses. 

Known issues: 

Some themes expose the fact that: 
- The background for the extension shelf when in detached mode (and located on the bottom) does not seamlessly blend in with background of new tab page. Still works surprisingly well when it breaks, though. 
- Didn't spend much time theming the shelf handle (just used the solid color from the theme). 
- When extension shelf is merged with the bookmark bar, there is a painting artifact (horizontal dotted line) that appears (apparently it was known to appear before this change).

BUG=18452, 21272, 21273
TEST=Install a theme for Chrome and make sure everything looks correct and is updated on a theme change. Also make sure painting problems in bugs 21272 and 21273 are fixed.

Review URL: http://codereview.chromium.org/204022

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26227 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 298e57790541914a60d0db4c6ecabc33104173be