Size the buffer used for the metrics log text properly.
The existing code was abusing the WriteInto interface (base/string_util.h).
The size passed to WriteInto should be the size of the string plus one for
a terminating NUL byte.

This bug caused the final newline in the log to be clipped.  c_str users
(there probably weren't any) would see the final newline which wouldn't be
wouldn't be followed by a NUL byte.

TEST=not really any
BUG=21733
Review URL: http://codereview.chromium.org/194099

git-svn-id: http://src.chromium.org/svn/trunk/src@26280 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2869a8c963b63c7d53b3f8e7c1caa75eaca217cd