Expectations for 3 perf systems on morejs and startup.
Set the perf expectations on linux, mac, and windows
perf systems for the morejs and warm startup tests.
The variance assumptions might require fine-tuning on
the master to get right.  We might be flaky to begin
with, or may want to turn off the variance reporting
altogether.

BUG=18597
TEST=build.chromium.org perf slaves stay green
Review URL: http://codereview.chromium.org/193083

git-svn-id: http://src.chromium.org/svn/trunk/src@26376 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 487f4f4d6002aaaece83275f9fcbee6a1a34ef30