Changing command buffer to use structs instead of arrays, step 1
I hope I'm not stepping on any toes here. I wanted
to write some command buffer code but I didn't want
to go poking values into arrays so I thought I'd
attempt to refactor a little. I hope I'm not steping
on any toes.

Some answers to possible questions.

O3D_COMMAND_ARGS_FLAG_AT_LEAST_N is a macro because
it needs to be a compile time constant so I can generate
the table.

All the commands with a FixMe are to be done in another
CL.

The Structs don't follow the style guide because it's
just busy work to match up SET_TOKEN and SetToken.
It seems like a reasonable exception to me. This way
the functions, IDs and structs all match easily
and therefore lists that use them be automatically
generated and kept in sync.

The next step is I want to uncompress some of the
commands. As an example, it doesn't seem like
CreateTexture2D should squeeze width and height into
16bits halfs of a 32bit value. This API is not targeted
at machines with 2K of memory and the amount of code
and other scaffolding around decoding those 2 16bit
halfs far outways saving 16bits in a command that
allocates a texture. It also makes it harder to use
the command because the person generating it has to
also go through the same hoops.


Review URL: http://codereview.chromium.org/206020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26421 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ee564a79995d8e91b7ddcbcd8f5b645b818b0803