Task Manager for OS X: Set up columns, show rows.
This is more work-in-progress for the OS X task manager. The table header is complete (the right columns are there, and right-clicking the header gives a column-selection context menu just like in e.g. Task Manager), and the rows data is displayed. However, all the metrics (memory, cpu, etc) seem to be all wrong. Because of this and because too much stuff is still unimplemented (favicons in rows, the "kill process" button, the link, etc), I won't commit the change to MainMenu.xib, so users still won't be able to open this dialog.

BUG=13156
TEST=Connect view->task manager to first responder's commandDispatch:, set the menu item to "enabled", and give the menu item the tag 40005. Now opening the task manager should actually display data. Right-clicking the table header should show a functional menu that lets you select which columns to display. Clicking a column header should not select the column, double-clicking an item should not edit it. Also, when resizing the task manager, stuff should now behave in a sane way.

Review URL: http://codereview.chromium.org/208043

git-svn-id: http://src.chromium.org/svn/trunk/src@26687 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c60ee852cd290a6c7d661a6275a741a74806cb81