Enable LeakTracker on release builds.
This is intended to be temporary, so we can get data from the dev channel on why URLRequest objects are leaking at shutdown.

WARNING: it may cause a slight slowdown on page cycler.

Also as part of this change, I switched the order that we check URLFetcher / URLRequest.
This is simply so that if both URLFetcher and URLRequest have leaked, we will report the URLFetcher leak rather than the URLRequest leak.

BUG=http://crbug.com/21199, http://crbug.com/18372

Review URL: http://codereview.chromium.org/217005

git-svn-id: http://src.chromium.org/svn/trunk/src@26765 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6b7b83aa5450730c470bcc3a6bbc91693002222d