Fix appcache_service and request_context referencing.
There is one appcache service per profile and several request context per profile. The profile holds a reference to the appcache service. Those request contexts which are subject to retrieval from appcaches hold a reference to the appcache service too. The appcache service is provided with a pointer back to the 'main' request context, this context is used when updating appcaches.

Initialization is a little tricky because profiles can't be used on the IO thread and request contexts can't be used on the UI thread.

BUG=22597,22125
TEST=many existing tests exercise profile/context creation
Review URL: http://codereview.chromium.org/215024

git-svn-id: http://src.chromium.org/svn/trunk/src@26844 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5dc508a45a7259be513bfe9304d0a64f18af3ab4