Fix the docs.  Apparently they've been broken since August 3rd rev 22348.
The scary thing is that's a little hard to believe
it's been broken that long and no one noticed
so I hope this is the correct fix but it does
appear to be wrong. The '/'s that were removed
are used in a string concat in the HTML templates.

Sometimes they need to be '' and sometimes 'folder/'
so that urls become

http://foo/bar.html

or

http://foo/folder/bar.html



Review URL: http://codereview.chromium.org/213033

git-svn-id: http://src.chromium.org/svn/trunk/src@26879 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 74cc62affd516d66af06cef9c14f41448522f461