Implemented upload progress indicator in status bar.
When sending an HTTP POST request with sufficiently large payload data (such as
uploading a file or several), the status bar should display the progress in the
form of a percentage. The ResourceDispatcherHost periodically sends upload
progress messages to the UI thread as well as to the renderer. If an upload is
occurring, other potential LoadStateChanged messages to the same RenderViewHost
are ignored; this prevents the upload progress display from being interrupted
periodically by "Waiting for <site>" messages. 

Original code review:
http://codereview.chromium.org/211050

Original patch by Simon Radford
r=darin
BUG=4715
TEST=Try uploading one or more files (e.g. at a file hosting site) and confirm
that the percentage is shown in the status bar. 

Review URL: http://codereview.chromium.org/244001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@27169 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 029d7cb2ce796ccee20d5ebd19989b53bba7d1d2