Make --show-extensions-on-top look exactly, pixel-for-pixel identical to the bookmark bar.
Used the same semi-transparent image for the button background
that the bookmarkbar uses, so that it will work for themes.

The upside of this is that we look the same as the bookmark bar. The downside is that we look the same as the bookmark bar (which currently looks pretty bad with some themes).

However, with the default theme, I think that the bookmark bar currently looks better than our toolstrips. The UI team should probably create some images that look good with different themes for the bookmark bar, then we can update this again.

Review URL: http://codereview.chromium.org/225025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@27258 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f6566a7e470c4e3127b31ddfaf8fd9dd60746383