Fix a bug where we get an extra resize event for background tabs.
The TabContentsContainerGtk is a GtkFixed that contains a
TabContentsViewGtk for each tab.  We hide the background TCVG
and only show the foreground TCVG.  However, when the TCCG
gets a resize event, we were still sending a set_size_request
to all the background tabs.  Since they were hidden, nothing
happened except in the case described in issue 14114.  It
seems like this is a spurious resize by gtk, but not sending
the resize request seems to work around it.

BUG=14114
TEST=see bug

Review URL: http://codereview.chromium.org/246012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@27297 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 03298f84381ac59a643edb4bb46716483dd7552b