Linux: add note to about:memoy re missing renderers.
On Linux, we cannot gets at the 'maps' nor 'smaps' files for our
renderers when they are sandboxed (because they are non-dumpable). One
could imaging plumbing IPC calls to the renderers for this
information, but it would be a lot of code for little gain. In the
mean time we should at least explain why all the renderers might be
missing.

(This is an issue with the SUID sandbox. The seccomp sandbox would be
ok.)

BUG=23258
http://codereview.chromium.org/235008


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@27388 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4aefa590cb7dc1ad085df1dfd5c309f60cdfbb5c