Quick fix for skinning translation bug.
We weren't binding the transform of the skin to the base
so the transform was getting applied twice and also
the math was wrong in skinning.

This is the shortest change to make this work but it
is arguably not the *correct* change.

The correct change would be to make the import code
create Skin separate from ParamArray and SkinEval
so that you can use the same Skin data multiple
times in the same scene. I don't know enough about
collada to know if that's possible to setup in
collada but the way the code is structured currently
it seems to assume it is even though the code does
support this correctly.

That would take some work to fix and require samples
and it seems unlikely people will run into that issue
so this smaller fix seems good enough for now.

Review URL: http://codereview.chromium.org/235046

git-svn-id: http://src.chromium.org/svn/trunk/src@27447 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a34cc813f72d10be891ceac7c080a31527a55477