Another attempt at landing this.
If IE was navigated away to a different page when the ChromeFrame inspector window was open
it would crash the browser.

The fix is to ensure that the inspector window is closed when the ExternalTabContainer window
is uninitialized.

This fixes bug http://code.google.com/p/chromium/issues/detail?id=22993

Bug=22993

Review URL: http://codereview.chromium.org/249028

git-svn-id: http://src.chromium.org/svn/trunk/src@27463 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7b6701ed3754bf278c83ea6a5d80fd47f1b33d5f