Another attempt at landing this.
If IE was navigated away to a different page when the ChromeFrame inspector window was open
it would crash the browser.

The fix is to ensure that the inspector window is closed when the ExternalTabContainer window
is uninitialized.

This fixes bug http://code.google.com/p/chromium/issues/detail?id=22993

Bug=22993

Review URL: http://codereview.chromium.org/249028

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@27463 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fc63434cf613554e299fe1c0554f19db3bab3e51