Fix cmd-up/cmd-down.
The approach is to special-case moveToBeginningOfDocument and moveToEndOfDocument in WebFrameImpl::executeCommand(). With this (and the cocoa keyboard bindings patch being landed), the special-case code in WebViewImpl::ScrollViewWithKeyboard() can be removed.

Also add a test for cmd-up/down.

Change chrome.gyp so that it supports osx-only unit tests (_unittest_mac.mm).

Move OnPrintPageAsBitmap to the other printing tests.

BUG=22585
TEST=Go to a page with both text box and scrollbar (e.g. http://en.wikipedia.org ). Pressing cmd-down should scroll to end of page, cmd-up back to start. Clicking the text box and trying some emacs shortcuts should work (ctrl-a jumps to start of line, cmd-h acts as backspace, etc).

Review URL: http://codereview.chromium.org/254002

git-svn-id: http://src.chromium.org/svn/trunk/src@27464 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7422d95ad9df857584a983e4ea3f138bd20d03c9