Fix issue #23442 - when test shell is run in layout test mode and a URL is passed in on the command line (which happens when running under Purify using --run-singly), the URL is canonicalized and echoed back into the output.
The net result of this is that if we passed in a file:///c:/ URL with a 
lower case drive letter, the output gets echoed back with an upper case
drive letter (file:///C:/). The test harness checks the output URL against
the input URL to make sure we're getting the output we expect, and in this
case it bails out (incorrectly).

To fix this, we change test_shell to echo out the URL as passed in.

  BUG=http://crbug.com/23442
  TEST=none
  R=darin@chromium.org

Review URL: http://codereview.chromium.org/249045

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@27681 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 25d5a238953b54315c99ff33e5613312fe165164