Relanding focus traversal on Linux toolkit views. It was previously breaking the interactive UI tests.
That was because I changed the button base class to be focusable by default which led to many button beings added to the tab traversal order.
I now explicitly make these buttons non focusable.

Also because of the way we clear the native focus on Gtk (by setting focus to NULL on the top level window), views now default to clearing the focus when focused (so that views non backed by a native window don't leave the native focus on the previously native window).
On Windows we used to focus the WidgetWin's HWND, and now we focus the top-level HWND. The WidgetWin class had to be changed to forward the key events to the root view that contains the focused view because of that. (otherwise it would send it to the top-level root-view, and if the focused view was in a child root-view it would not recieve the keyboard messages).
Note that the mouse wheel does not need that, as the mouse-wheel Windows messages are sent to the HWND under the cursor.

See:
http://codereview.chromium.org/246030/show

TEST=Run the tests
BUG=NONE

Review URL: http://codereview.chromium.org/255008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@27723 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ad3e3faffb49cced19aef660888baa92567eeea6