Add some initial code for the top sites service. This will be a replacement for the thumbnail database, and will also replace the ThumbnailStore which was the previous replacement we didn't ship. This component will be very much like the ThumbnailStore wtih the addition of the actual most visited data (not just thumbnails) and that it is threadsafe.
This class is designed to be called on any thread. When it is complete,
thumbnails will be added to it from the UI thread of the browser. Requests for
thumbnails and the most visited data can be serviced directly on the I/O thread
without going through the UI thread, and since the data is cached, the request
won't also have to go through the history thread.

The current state is that it cqan store and update the the most visited list and thumbnails. There are unit tests covering this behavior.

I also added support for redirect ranking to ThumbnailScore. This is to
duplicated the ranking function in history currently, where it prefers
thumbnails closer to the end of a redirect chain. Since we won't be using the
history service and are only storing thumbnails for the most visited items, we
have to track the redirect index ourselves.

BUG=none
TEST=covered by unit tests (hopefully!)
Review URL: http://codereview.chromium.org/251002

git-svn-id: http://src.chromium.org/svn/trunk/src@27824 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 01edf903205d8da41d02997e8787efd4b0c89c78