This makes some changes to the o3d tree in preparation for landing in Chrome.
The biggest pieces here are moving some of the third party
dependencies back into o3d/third_party because they need
to be distinct from the chrome versions of the same packages,
and because O3D is the only one using the dependency.
(NPAPI in particular).

Also the plugin gyp file is now modified so that it can handle
being part of both a chrome developer tree and an o3d developer
tree (in the latter case, it generates the plugin and installer, in
the former it does not)

BE AWARE that this change will change the main solution/xcodebuild file
from "build/all" to be "build/o3d", but rebuilding from gyp files will NOT
remove the old "build/all" one, so you might be tricked into opening the
wrong one.

Review URL: http://codereview.chromium.org/256081

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28169 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9641ce879ffc420036cceff15b4e3f886a694e56