GTK: Draw bottom-aligned ntp themes on floating bookmark bar.
This is a simplified version of the windows code. There will be a follow-up patch to switch windows over to using this as well.

I ran into some trouble calculating the correct size of the tab contents. Punting on that for now, filed http://crbug.com/23907

BUG=22836
TEST=Tested on klassen, candies, karim rashid, tiesto, ratchet and clank, and jeff coons themes.

Review URL: http://codereview.chromium.org/255086

git-svn-id: http://src.chromium.org/svn/trunk/src@28215 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 87223cb7624333d5cd40c7ae21d3fbad658e5e56