Fix --new-baseline to honor the --platform flag.
Also test_types/text_diff.py attempts to be smarter than the user by checking to see if the text results are a render tree output and, if not, writing them into the generic directories rather than the specified directory. Except, that code looks broken, too. I think it's probably best to just do what the user told it to do, so I've removed that path.

Thanks to senorblanco for finding the bug.

  R=ojan@chromium.org
  TEST=none
  BUG=none

Review URL: http://codereview.chromium.org/255068

git-svn-id: http://src.chromium.org/svn/trunk/src@28220 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 708cabbbf036068c73296a468cc298906adfaa4a