Add about:memory link to task manager.
Since about:memory DCHECK()s atm, the link is not completely hooked up.

In the xib, I've added a square button with height 14 and the new HyperlinkCell, contained in a GTMWidthBasedResizer or how it's called. I made sure the baseline of link on the left and button on the right is at the same height.

BUG=17989,13156
TEST=Open MainMenu.xib, click "Enable" for view->dev->taskman. Build&run chrome. Task manager should contain link.

Review URL: http://codereview.chromium.org/255018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28426 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7b70d383240efd3282392c893c4c0ad689184554