Get rid of mac_app_names.h.  Replace it with real constants from chrome_constants.h.
As a side effect of this change, tests will use the framework in the build
directory as their main bundle override for the purposes of resource loading,
instead of looking into the framework inside the application bundle.  This
should be more than sufficient, and is done for simplicity, because a future
change will make it more difficult to locate the correct framework inside the
application bundle unless the product version number is known.

BUG=24220
TEST=product still works and all tests still pass
Review URL: http://codereview.chromium.org/266023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28431 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9a42779a1b6252e64da89a6e9166e08589a98da9