Disk cache: Add a method to cancel pending sparse operations.
The sparse IO methods require exclusive use of the cache entry
and they complain when that requirement is violated. When the
user cancels a request and reissues another one to the same entry,
we may be waiting for the previous operation to finish when we
receive a new IO request, so we fail.

This CL add a way for the HTTP cache to cancel IO operations and
get a notification when the disk cache is able to operate on that
entry again.

BUG=23862
TEST=unittests

Review URL: http://codereview.chromium.org/256090

git-svn-id: http://src.chromium.org/svn/trunk/src@28475 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d709b91a4bb907639a006ab589f743e0cdc7d3f3