Revert "Revert "Start the database query for the new tab page results as soon""
This re-applies r28549 with a NULL pointer check for the HistoryService, which
can be NULL during unit tests.

It also means that if the user does a view-source on the new tab page, we make
a wasted query to the history service, but I guess that's not a big deal.

TBR=arv

Review URL: http://codereview.chromium.org/267039

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28559 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f5aa8ee6979a91c9f0394a90b27c38e62ec4a2c2