Revert 28558 because this is clearly responsible for breaking the interactive ui tests. It just needs a clobber.
It looks like 2 parts of the code don't agree on the number of
buttons on the bookmark bar. Maybe because part of the code is
compiled with CHROME_PERSONALIZATION = 1 while the test is compiled
with CHROME_PERSONALIZATION = 0. Maybe this new gyp change broke the
define propagation. This is really easy to replicate, and should also
be easy to debug.

Original Log:
Relanding the interactive UI tests GYP factor out, as it turns
out it is not responsible for breaking the interactive ui tests.


Review URL: http://codereview.chromium.org/271046

git-svn-id: http://src.chromium.org/svn/trunk/src@28668 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 48999ebb183ac4b3eb3ca8a7378de65ad3bf186d