Implement a version of NativeMetafile for the Mac
This is the first of series of patches to implement printing on the Mac.

This API is based heavily on the Linux version; although it doesn't end up abstracting very much on the Mac, having a NativeMetafile implementation makes it easy to pass data through the existing print architecture.

BUG=13158
TEST=Printing unit tests; no effect in the application yet.

Review URL: http://codereview.chromium.org/269041

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28722 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2bb59f2ee31500cc9b25838a2ea711e3c7d5e6fd