More gypbuild infrustructure.
No idea if this is the best structure or the correct
options but I have to start somewhere. If you have
better ideas speakup or start editing :-D

If there is a way to get gpy to list the targets
please tell me so I can call it.

Review URL: http://codereview.chromium.org/242068

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28744 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8dc868f5d408b862cf347b9498182dd3c1212a3b