More gypbuild infrustructure.
No idea if this is the best structure or the correct
options but I have to start somewhere. If you have
better ideas speakup or start editing :-D

If there is a way to get gpy to list the targets
please tell me so I can call it.

Review URL: http://codereview.chromium.org/242068

git-svn-id: http://src.chromium.org/svn/trunk/src@28744 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: faa3da1f112e09f79d1a23af70350b19debea9d2