GTK: avoid a hang brought on by an infinite size-allocate queue.
This is not the most satisfying fix imaginable. I'm not sure why we are getting size-allocate events where the visibility of the chevron has changed but our allocation is not yet updated. In principle I suppose it would be better not to call any function that might cause an allocation from within a size-allocate handler, but guaranteeing that is probably pretty hard.

BUG=24470
TEST=repro steps no longer repro; also, verified that the early return line is actually being hit.

Review URL: http://codereview.chromium.org/270078

git-svn-id: http://src.chromium.org/svn/trunk/src@28778 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3f08ab8c954b36279658b5c6c6d9fd68806940c4