GTK: avoid a hang brought on by an infinite size-allocate queue.
This is not the most satisfying fix imaginable. I'm not sure why we are getting size-allocate events where the visibility of the chevron has changed but our allocation is not yet updated. In principle I suppose it would be better not to call any function that might cause an allocation from within a size-allocate handler, but guaranteeing that is probably pretty hard.

BUG=24470
TEST=repro steps no longer repro; also, verified that the early return line is actually being hit.

Review URL: http://codereview.chromium.org/270078

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28778 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: eda5115bfb267e40790137ac4fe22e13724c3c3c