For sync exponential backoff, allow one nudge per exponential backoff interval. If the nudge still leaves the syncer with more work to do, don't accept any further nudges for this interval, and keep the exponent stage for exponential backoff the same.
As a result of the patch, the unittest can now explicitly determine if exponential backoff
kicked in or not.

We really need to wire this up directly to the error codes to be precise, because the current impl (and
it's the same with my patch) appears to trigger exponential backoff in other cases (ShouldSyncAgain).

TEST=SyncerThreadTest

Review URL: http://codereview.chromium.org/275015

git-svn-id: http://src.chromium.org/svn/trunk/src@29053 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6e51f1cb1b99ab3b1f3bb4de939a641df4e9f8a6