Add env variable to write minidumps to a preset directory.
Add an |BREAKPAD_DUMP_LOCATION| environmental variable to allow overriding the default location
for minidump generation.

This is for ChromeBot who need a single copy of the Chrome bundle which can be run several times
alternating the location of crash dumps.

This isn't implemented as a command line flag because only some command line flags are passed to child processes and we need a robust solution.

TEST=Minidumps should still be written for browser and renderer crashes.

Review URL: http://codereview.chromium.org/264038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@29126 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 913241aaa693765c71df716f6e4e041532f484b5