Removing a synchronous pause that reduces impact on the IO thread
Audio pause is implemented synchronously in the browser process.
This can easily choke the IO thread when user does a series of
pause of looping a very short audio / video. This change will
make the pause action asynchronous. The drawback of this change
is that we cannot stop audio playback immediately as we receive
the pause request.

This is still an improvement over chrome 3.0 as the "playback"
period after pause is noticably shorter.

Review URL: http://codereview.chromium.org/279010

git-svn-id: http://src.chromium.org/svn/trunk/src@29234 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5bf654f0875ad8359c65cd04374b64e85b7fec9d