Remove an overzealous version check from keystone_install.sh.
The existing version check compared the previously-installed version's
KSVersion with the newly-installed version's KSVersion after completing the
update.  In a case where a previous update attempt was interrupted after
the new KSVersion was set, this check could fail on a subsequent run of the
update script.

BUG=none
TEST=keystone_install_test.sh
Review URL: http://codereview.chromium.org/287006

git-svn-id: http://src.chromium.org/svn/trunk/src@29296 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ff23ec2f29b0a53b6fa50c47b9ac0da1ed3da01a