Use a multiset rather than a set to store bookmarks when running the model associator merge.
For cases where there's an exact duplicate URL and title,
this allows each BookmarkNode a chance to be associated
with a sync node.  With a set, we'd forget about all
but one of the duplicates, resulting in the creation
of yet more duplicates.  The result was exponential
growth with each merge and sync.

BUG=24995
TEST=

Review URL: http://codereview.chromium.org/288001

git-svn-id: http://src.chromium.org/svn/trunk/src@29307 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ede8751cef7090a43de63beb28298e3890e2fe31