Classify more types of input as queries.
This adds an oracle that can determine if a hostname is RFC 1738 compliant; if we get a hostname that's not an IP address and not compliant here, we almost certainly have a query.

This also rejects "ports" that don't consist of a number between 0 and 65535 (e.g. port "abc" or port "123456").

BUG=18817
TEST=The string "foo: bar" (with quotes) should default to search, not navigate, in the address bar.
Review URL: http://codereview.chromium.org/271116

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@29317 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1e9943abd4d5b08bd03f813661f8e6cb820f7982