Fix ProfileSyncService unit test.
Create TestHttpBridge for use by ProfileSyncServiceUnittest, so it
doesn't crash.

DoInitialize has been changed to take a HttpPostProviderFactory
before the core initialization.  The HttpBridge is now created
on the UI thread instead of the core thread.

Create fake tagged nodes in the order that they are requested.

Make the request order consistent and in visual order since this is the
most natural way to request them.

TEST=Ran the affected unit tests on Windows, tested basic sync functionality with two concurrent browsers.
BUG=25021

Review URL: http://codereview.chromium.org/292008

git-svn-id: http://src.chromium.org/svn/trunk/src@29368 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5786b39ee48bc5268de626da822b8c007b95186d