When Chrome hands off a URL to be opened by the external host by the ViewHostMsg_OpenURL IPC, it needs to pass the referrer as well. The Chrome fixes in this CL are mostly related to passing the HTTP referer off to the browser and from there to the ExternalTabContainer to ChromeFrame and back.
The ChromeFrame changes are basically around the same lines with one exception. When we handle the
AutomationMsg_OpenURL IPC in the activex and the active document we pass the referer if applicable
to the WebBrowser2::Navigate2 interface, which is then read by the BHO in BeforeNavigate2. We then
save away an AddRef'ed BHO pointer in TLS which is then referenced by the Active document for reading
the referer and passing it off to Chrome in the NavigateInExternalTab message.

Added a unit test in ChromeFrame which tests this case.

This fixes http://code.google.com/p/chromium/issues/detail?id=22994

Bug=22994

Review URL: http://codereview.chromium.org/274071

git-svn-id: http://src.chromium.org/svn/trunk/src@29420 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0de95d1213b39136e3b3652f50da8a3eb0674bf3