Pull in googleurl @ r=120 to fix the query part encoding handling (the patch by Roland Steiner).
What Roland wrote at http://codereview.chromium.org/243028/show

Encoding of the query part of an URL was inconsistent: 
If it was part of a larger relative URL, it would get encoded. 
However, if the query part was stand-alone, then it would not be encoded. 

As indicated by the original layout test referenced in the bug (see below), 
this patch changes the code so that the query part is encoded in both cases. 

Note, however, that with this patch the layout test will NOT yet fully pass, 
since it also handles the fragment part differently. 
As the different handling of the fragment part seems to be quite on purpose, 
I have not modified the behavior - see the discussion at 

http://code.google.com/p/chromium/issues/detail?id=20507 . 


BUG=8912,20507
TBR=brettw


Review URL: http://codereview.chromium.org/297010

git-svn-id: http://src.chromium.org/svn/trunk/src@29470 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 57055700735dc72711fb9a40af430ee57d87d873