1. Make messageloop-based timeout work on Mac/Linux, 2. Save the results of timing out tests and show them in the results.html file.
One arguably negative side effect of this is that a test that times out and has
image/text failures will count as both a timeout failure and an image/text failure.

I'm not 100% certain the RevokeAll change makes sense. It matches what the old
code's comments claimed it did, but the code didn't actually do. The idea is
that if the test finishes before the timeout is hit, we should cancel the timeout.
Review URL: http://codereview.chromium.org/303007

git-svn-id: http://src.chromium.org/svn/trunk/src@29576 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3d459e747ccb467539c8726216eabe3dc94737e8