Add a mechanism to disable IPv6.
(1) Adds the ability to specify the address family on a per-request basis.

(2) Exposes a --disable-ipv6 flag to chrome that changes the default address family from AF_UNSPEC to AF_INET (same sort of thing Firefox does).

(3) Changes the backing datastructure for HostCache:EntryMap and HostResolverImpl::JobMap from a "hash_map" to a "std::map". This was for consistency with other code (when I went to add a custom hash trait, I couldn't find any existing code which was using hashmap for custom keys).

(4) Updates about:net-internals to display an address family for the hostcache dump (since it is now a part of the key).

This change is in anticipation of turning off IPv6 host resolving in the PAC utility functions (see bug 24641). But it is also a feature addition.

BUG=24641

TEST=HostCacheTest.AddressFamilyIsPartOfKey
Review URL: http://codereview.chromium.org/302010

git-svn-id: http://src.chromium.org/svn/trunk/src@29686 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d06bf3419d30bf3e0f671b3eefb847e3a25e86eb