Fix for issue 24888
Richedit seems to have a default paragraph format which will be used when the last paragraph in the document is deleted. (e.g. when SetWindowText is called, or when user manually deletes everything). When this happens, richedit also automatically set keyboard layout to match the default paragraph format. Unfortunately, fix for issue 11683 does not help for this scenario. The default paragraph format seems could only be set when user switches keyboard layout with an empty richedit. The fix in this CL works around the problem by setting richedit's text to empty before it forward the reading order change hotkey (ctrl+shift/shift+ctrl).

BUG=24888
TEST=Verify this CL fixes issue 24888 that LTR/RTL reading order is preserved when user selects suggestions/changes tabs after changing keyboard layout with non-empty text in Omnibox.

Review URL: http://codereview.chromium.org/294013

git-svn-id: http://src.chromium.org/svn/trunk/src@29695 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: fae65338eff015f45797c988479fc830122791c1