GCC's optimizer is getting more aggressive. It is no longer good enough to just pass the address of a structure as an input parameter to assembly code. The assembly code must also mark "memory" as getting clobbered, even if it only wants to read from the structure. This seems to be a result of strict aliasing and the lack of an ability for the assembly code to clearly say which pointers it dereferences.
Furthermore, if the assembly code touches the stack (e.g. uses
"push"), it must now mark the stack pointer as getting
clobbered. Otherwise, GCC assumes that the red zone won't be
clobbered, and that it is possible to use the stack pointer as an
input register.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/320008

git-svn-id: http://src.chromium.org/svn/trunk/src@29829 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 08e87ff133940813d3824152bf63ba05574ae775