Fixes to get renderer=cb cb_service=gl to compile again.
I was working on spliting some of the command buffer 
code in preparation for GL command buffers so those
changes are in there as well. I didn't realize the
gl command buffers were broken.

I'm planning on sharing some of the command buffer
code and commands so I thought about making the
command buffers into 2 parts. The first N (256?)
commands are common commands.  Noop, SetToken, Jump
Gosub. Those will be handled by common code.

After that come the O3D or GL command buffer
commands. I'm not sure that's a good idea but
there is a significant amount of code in managing
tokens and parsing etc and it seems like it would
be nice to share it until/if we delete the O3D 
command buffers.

Note: You'll see that SetToken is hardcoded to 1.
I plan to fix that in the next CL when I split
the command into shared vs specific.

Review URL: http://codereview.chromium.org/307043

git-svn-id: http://src.chromium.org/svn/trunk/src@29831 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c1775b12e2d29a319fbe647eb0007f2a1ad0369f