Potential fix for the PluginChannel::CleanUp crash.  This will go out on tomorrow's dev channel build and we can see if the crashes go away while I try to write a repro.
The only scenario that I can think of is that an NPObject in the plugin process has a deallocate function which releases an npobject from the renderer.  That would cause the corresponding NPObjectProxy to go away, but since PluginChannel has in_remove_route_ set, the list won't get updated and in the future iteration of the loop we'll call the method on a deleted object.

BUG=25439
Review URL: http://codereview.chromium.org/327003

git-svn-id: http://src.chromium.org/svn/trunk/src@29839 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1eb9d17ced10f689db428f7b02b9b4a3a544d6ca