Submitting on behalf of jschuh@ See http://codereview.chromium.org/307020
Changed active URL handling in RenderView::OnMessageReceived and RenderView::OnNavigate. Now we call SetActiveURL directly so that we don't unset the URL when falling out of scope. 

We can try this in the render process on dev channel for a few weeks. If we're happy with the results, we do the same thing on the plugin process. 


TBR=cpu
BUG=22033
TEST=Trigger a crash in the render process and see if the dump includes the URL

Review URL: http://codereview.chromium.org/328008

git-svn-id: http://src.chromium.org/svn/trunk/src@29933 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 9b44f8694f7f94b8d57bc499a5b11be82b7097ed