Mac: implement dispositions for back/forward buttons (not in drop-downs).
That is, holding Cmd while pressing back/forward will open in a new background tab and holding Cmd+Shift will open in a new foreground tab. Theoretically, it can support holding Shift to open in a new window, but Browser::GoBack()/GoForward() doesn't support this disposition (due to a technical "limitation" of the implementation?).

This does *not* implement dispositions in the drop-down menus.

BUG=25319
TEST=Navigate to build up some history. Then press back/forward while holding down Cmd or Cmd+Shift.

Review URL: http://codereview.chromium.org/330002

git-svn-id: http://src.chromium.org/svn/trunk/src@30016 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d77d6dcf09c6b052051524b1fa8de8fc490380b3