Do not send some keyboard shortcuts to the renderers
Walking the whole menu on every keypress seems ridiculous. Linux does this too :-/ (Caching is made hard because the user can change key equivalents in system preferences at every point in time, and we're not notified of that. And people only hit max 5 keys/second, so it's not all that ridiculous).

There's a UI test for this, but the interactive UI tests are not enabled on OS X, so it's not executed.

Menu walking code based on code from CocoatechCore.

BUG=5496,15090,24877
TEST=Go to http://unixpapa.com/js/testkey.html , check "keydown", focus the textbox, and make sure that cmd-t still opens tabs, cmd-shift-[ still switches tabs, cmd-w still closes tabs. however, cmd-L should not focus url bar and cmd-1 should not go to the first tab.

Review URL: http://codereview.chromium.org/273032

git-svn-id: http://src.chromium.org/svn/trunk/src@30032 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4abe1b4233d793baf861d74d2e4ae69b9d019d56