cros: doing dbus stuff on the file thread==disaster.
libdbus-glib uses the glib main loop internally.  We were making calls on a background thread that wound up telling libdbus-glib to get a connection to the system bus on said background thread.  This led to a background thread trying to use the same glib main loop as Chrome's UI thread, and concurrency issues led to crashing.  Sadly, there can be only one main loop per process, as I understand it.

This code takes away the attempts to use the background thread.  As our interaction with dbus is semantically asynchronous, this seems ok to me. 
Review URL: http://codereview.chromium.org/339013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30135 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3fe9757ece0b3cc7e5746716608554862118aa56