CL which includes http://codereview.chromium.org/337006 for mrossetti Also includes a unit test fix. BUG=21908
TEST=Control-click on the bookmark bar, either on an existing folder
or on the blank area and choose "Add Folder..." from the contextual
menu. A sheet will drop down from the bookmark bar asking for the name
of the folder to add. The edit field will be populated with the
proposed name 'New folder'. ('New folder' will be appropriately
localized.)


Review URL: http://codereview.chromium.org/329031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30207 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 500df63b7cb64beffe44e9712aa20d20703e1de4