Get rid of FilePath::AppendAndResolveRelative().
To resolve the problem of '..' parent references as well as symbolic links on POSIX platforms, we can simply use the file_util::AbsolutePath() function.

This has the drawback of having a different behavior on Windows and POSIX platforms, in the way that it can return a canonical path that doesn't exists when ran on Windows, but it will return an empty path (or false) when run on a POSIX platform.

So we need to add an extra PathExists() call to unify the behavior.

BUG=25681,25131 

Review URL: http://codereview.chromium.org/343003

git-svn-id: http://src.chromium.org/svn/trunk/src@30334 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c3160aa0125cff7ff025801305bd34db56108f13