Use a regular interface pointer for ITextDocument in AutocompleteEditViewWin.
I'm doing this because we want precise control over the lifetime of the validity of this pointer.

It turns out the paint patching code caused richedit20.dll to be unloaded, which caused the pointer held by the ScopedComPtr member to become junk and the subsequent ~ScopedComPtr to crash.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/340014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30429 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bbae713567621f650f8e8b8125cb3f5ca762d2b5