Use a regular interface pointer for ITextDocument in AutocompleteEditViewWin.
I'm doing this because we want precise control over the lifetime of the validity of this pointer.

It turns out the paint patching code caused richedit20.dll to be unloaded, which caused the pointer held by the ScopedComPtr member to become junk and the subsequent ~ScopedComPtr to crash.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/340014

git-svn-id: http://src.chromium.org/svn/trunk/src@30429 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5c61d4c235211d8a3fb44dc27ae792d45c117c72