Fix transparent windowless plugin background painting. This would manifest itself as incorrect ClearType edges.
The problem was that we were taking into account the translation from the page
coordinate system when blitting the background, but this had already been
accounted for in the transformation matrix, resulting in duplicate
transformations.

The fix simplifies the code by not tranlating into the page coordinate system
until later, staying in the plugin coordinate system which is more convenient.
This means that the background (which is the same size as the buffer) can just
be painted at (0,0) with no fuss and no muss. We don't even need to worry about
how much we paint since the clip was already installed.

This also corrects some indenting in browser.cc that I noticed.

TEST="Attach a file" text in Gmail compose looks weird (sometimes requires
     selecting it to make this happen).
BUG=26080
Review URL: http://codereview.chromium.org/343040

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30498 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7b20051ce2f6359a5d86b05d1af9f8e77434540b