Make LoadLog refcounted-thread safe.
This also backs out r26610 (these are the changes to host_resolver_impl.cc), since that code is no longer necessary if LoadLog is refcounted thread safe.

LoadLog itself is not thread safe -- however making its refcounting threadsafe simplifies passing it across thread boundaries.

Initially I had avoided doing this, and instead was trying to be very careful in when callers would AddRef/Release to avoid races.
This approach however turned out to be fragile, and it was difficult to chase down all of the dependencies (they just kept popping up).
So now I am changing that decision, and opting for the simpler approach which will be much more robust.

The performance impact should be negligable.

BUG=22272

Review URL: http://codereview.chromium.org/347021

git-svn-id: http://src.chromium.org/svn/trunk/src@30525 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 769e4ac69b89057e41f10187c8ffee76950afd25