Improve memwatch tool.
I removed a leak of "stack" objects, and made them a little
more lightweight by removing an unused member.

This change is only part of an option for chrome, that is not currently
linked in during any tests.  It can only currently be used under XP, so
build-bot tests are not currently appropriate.

r=mbelshe
Review URL: http://codereview.chromium.org/345033

git-svn-id: http://src.chromium.org/svn/trunk/src@30672 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2f31e1d77e9b22c3b96a80ec0e4ac3187d95b5a2