Better handling of UNKNOWN versus QUERY in a couple ways.
First, made it possible to navigate to "invalid" hostnames in a number of additional cases, including when the user types an explicit scheme or when we have a valid TLD and no spaces.  The default action for these kinds of inputs is still search, but we'll pull up an accidental search infobar if need be now.

Second, made the HistoryURLProvider show a What You Typed match for UNKNOWN inputs when the user explicitly typed "http:" or "https:" at the front, since these have the highest probability of being navigational inputs.

BUG=26341
TEST=Type "http://foo bar baz.com" and verify you are given a (non-default) entry in the dropdown to navigate.
Review URL: http://codereview.chromium.org/353010

git-svn-id: http://src.chromium.org/svn/trunk/src@30830 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e03fbb526b7be5d6ed79adef69e92ff628833ee2