Fullscreen Exit Bubble should not elide its label.
When drawing a Label it is important to adjust for
RTL not just when drawing, but also when computing
the size of the string.

Before this change, we'd compute using Font::GetStringWidth
(which uses ::GetTextExtentPoint32 *without* adjusting
the string for RTL) and draw using DoDrawText (which
adjusts for RTL and then draws using ::DrawText).

After this change, we now use SizeStringInt for
computing (so we now compute and draw using the
same ::DrawText function, thereby adjusting the 
string for RTL in both cases).

BUG=17949
TEST=Run chrome with --lang=he, press F11 and
the string in the bubble at the top should not
be elided.

Review URL: http://codereview.chromium.org/354014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30863 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2cc65abb75702400bd3ad90422a624e8b1f1d807