Fullscreen Exit Bubble should not elide its label.
When drawing a Label it is important to adjust for
RTL not just when drawing, but also when computing
the size of the string.

Before this change, we'd compute using Font::GetStringWidth
(which uses ::GetTextExtentPoint32 *without* adjusting
the string for RTL) and draw using DoDrawText (which
adjusts for RTL and then draws using ::DrawText).

After this change, we now use SizeStringInt for
computing (so we now compute and draw using the
same ::DrawText function, thereby adjusting the 
string for RTL in both cases).

BUG=17949
TEST=Run chrome with --lang=he, press F11 and
the string in the bubble at the top should not
be elided.

Review URL: http://codereview.chromium.org/354014

git-svn-id: http://src.chromium.org/svn/trunk/src@30863 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2193a03ba8b1ccb3248cead3a6565ab02760da8e