Attempt 2 at:
Gets find bar animation/clipping to work on views/gtk.

The only difference between this and the first version is fixing an
include in extension_host that was triggering mac to build views, as
well as an extraneous SetSlideDuration I had used for testing.

BUG=none
TEST=none
TBR=ben
Review URL: http://codereview.chromium.org/342116

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30942 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6b478019d0c242b0ddbca6366adf6e6e5e6278cd