Add default implementations for WebViewClient and WebWidgetClient methods.
Once these interfaces live upstream, we'll want to allow the addition of new methods w/o requiring the consumer (chrome) to implement the new methods.  That'll make webkit merging easier, and it'll also make it possible to use a newer webkit with an older chrome.  Obviously, this only works if the methods are not crucial to webkit working properly, but that is generally the case for *Client methods.

R=dglazkov
BUG=none
TEST=none


Review URL: http://codereview.chromium.org/346042

git-svn-id: http://src.chromium.org/svn/trunk/src@30943 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0495dfb8810ebd19906d5a66b3032e9f6a4abd95