Add default implementations for WebViewClient and WebWidgetClient methods.
Once these interfaces live upstream, we'll want to allow the addition of new methods w/o requiring the consumer (chrome) to implement the new methods.  That'll make webkit merging easier, and it'll also make it possible to use a newer webkit with an older chrome.  Obviously, this only works if the methods are not crucial to webkit working properly, but that is generally the case for *Client methods.

R=dglazkov
BUG=none
TEST=none


Review URL: http://codereview.chromium.org/346042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30943 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d54e84f06860d37e47499807e17be3d2e4546d7a