Mac: don't restore explicitly closed windows.
With "On startup"/"Restore the pages that were open last" set (in Preferences), a window which was explicitly closed (using the red button, Cmd-Shift-W, the menu item -- but not due to quit) should not be restored on startup. This is the behaviour of Firefox and Camino (Safari doesn't appear to implement this feature).

Note about this patch: Depending on how we decide things will work for App mode, much more code could be disabled on Mac.

BUG=13341
TEST=Set the browser to restore pages on start (see above). Navigate somewhere interesting; close window; quit; restart; should open a browser with default window (NTP or homepage or whatever). Navigate somewhere; quit; restart; should open previously opened stuff. Make sure that session restore (also crash recovery) still work as expected.

Review URL: http://codereview.chromium.org/362016

git-svn-id: http://src.chromium.org/svn/trunk/src@31164 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8fb7e3526d0260c05ace6019f273b004e9a3a61c